Refactored spectron tests
