+++ tests/clipboard.test.js
@@ -1,15 +1,13 @@
+const Application = require('./spectron/spectronSetup');
 const path = require('path');
+let app = new Application({});
 
 describe('Tests for clipboard', () => {
 
     let originalTimeout = jasmine.DEFAULT_TIMEOUT_INTERVAL;
-    jasmine.DEFAULT_TIMEOUT_INTERVAL = 90000;
-
-    let app;
+    jasmine.DEFAULT_TIMEOUT_INTERVAL = Application.getTimeOut();
 
     beforeAll((done) => {
-        const Application = require('./utils/spectronSetup');
-        app = new Application({});
         return app.startApplication().then((startedApp) => {
             app = startedApp;
             done();
